FEAT: upload capability implementation and test
